rename componenets, fix for empty device types list
